% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bioc_install_if_missing}
\alias{bioc_install_if_missing}
\title{Install Missing Bioconductor Packages}
\usage{
bioc_install_if_missing(packages)
}
\arguments{
\item{packages}{Character vector of Bioconductor package names}
}
\value{
None (invisible)
}
\description{
Helper function to install missing Bioconductor packages.
}
