% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_analysis.R
\name{convert_to_viper_format}
\alias{convert_to_viper_format}
\title{Convert to VIPER Format}
\usage{
convert_to_viper_format(regulon)
}
\arguments{
\item{regulon}{Regulon data frame with ENSEMBL IDs}
}
\value{
VIPER-formatted regulon list
}
\description{
Converts DoRothEA regulons to VIPER format.
}
\keyword{internal}
