% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincs_analysis.R
\name{create_lincs_connectivity_plot}
\alias{create_lincs_connectivity_plot}
\title{Create LINCS Connectivity Score Plot}
\usage{
create_lincs_connectivity_plot(
  results_table,
  output_dir,
  experiment_name,
  n_compounds = 30
)
}
\arguments{
\item{results_table}{LINCS results table}

\item{output_dir}{Output directory}

\item{experiment_name}{Experiment name}

\item{n_compounds}{Number of compounds to show (default: 30)}
}
\value{
ggplot object

ggplot object
}
\description{
Creates a plot showing connectivity scores for top compounds.
}
\keyword{Connectivity}
\keyword{Create}
\keyword{Creates}
\keyword{LINCS}
\keyword{Plot}
\keyword{Score}
\keyword{a}
\keyword{bar}
\keyword{compounds.}
\keyword{connectivity}
\keyword{for}
\keyword{internal}
\keyword{plot}
\keyword{scores}
\keyword{showing}
\keyword{top}
