% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincs_analysis.R
\name{create_lincs_query_signature}
\alias{create_lincs_query_signature}
\title{Create LINCS Query Signature}
\usage{
create_lincs_query_signature(up_genes, down_genes)
}
\arguments{
\item{up_genes}{Character vector of upregulated gene IDs}

\item{down_genes}{Character vector of downregulated gene IDs}
}
\value{
signatureSearch qSig object
}
\description{
Creates a query signature object for LINCS analysis.
}
\keyword{internal}
