% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_output_dir}
\alias{create_output_dir}
\title{Create Output Directory}
\usage{
create_output_dir(dir_path)
}
\arguments{
\item{dir_path}{Path to directory to create}
}
\value{
Character path to directory
}
\description{
Creates output directory if it doesn't exist.
}
