% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_analysis.R
\name{create_string_network}
\alias{create_string_network}
\title{Create STRING Network}
\usage{
create_string_network(
  gene_symbols,
  module,
  species_id,
  output_dir,
  experiment_name
)
}
\arguments{
\item{gene_symbols}{Gene symbols}

\item{module}{Module name}

\item{species_id}{STRING species ID}

\item{output_dir}{Output directory}

\item{experiment_name}{Experiment name}
}
\value{
STRING network data
}
\description{
Creates protein-protein interaction network using STRINGdb.
}
\keyword{internal}
