% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_analysis.R
\name{create_trait_matrix}
\alias{create_trait_matrix}
\title{Create Trait Matrix}
\usage{
create_trait_matrix(metadata, trait_column)
}
\arguments{
\item{metadata}{Sample metadata}

\item{trait_column}{Column name containing trait information}
}
\value{
Numeric matrix of traits
}
\description{
Converts trait information to numeric matrix for WGCNA analysis.
}
\keyword{internal}
