% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{filter_protein_coding_genes}
\alias{filter_protein_coding_genes}
\title{Filter for Protein-Coding Genes}
\usage{
filter_protein_coding_genes(expr_df, gtf)
}
\arguments{
\item{expr_df}{Data frame with gene expression (must have a 'gene_id' column)}

\item{gtf}{GTF object loaded with rtracklayer::import()}
}
\value{
Data frame with only protein-coding genes, rownames as geneName_geneID
}
\description{
Filters expression data to keep only protein-coding genes using GTF annotation.
}
