% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_required_bioc_packages}
\alias{get_required_bioc_packages}
\title{Get Default Bioconductor Packages}
\usage{
get_required_bioc_packages()
}
\value{
Character vector of Bioconductor package names
}
\description{
Returns the default list of required Bioconductor packages.
}
