% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_required_packages}
\alias{get_required_packages}
\title{Get Default Required Packages}
\usage{
get_required_packages()
}
\value{
Character vector of package names
}
\description{
Returns the default list of required packages for the pipeline.
}
