% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_analysis.R
\name{identify_significant_modules}
\alias{identify_significant_modules}
\title{Identify Significant Modules}
\usage{
identify_significant_modules(
  wgcna_results,
  significance_threshold,
  min_correlation_threshold
)
}
\arguments{
\item{wgcna_results}{WGCNA analysis results}

\item{significance_threshold}{P-value threshold}

\item{min_correlation_threshold}{Minimum correlation threshold}
}
\value{
Vector of significant module names
}
\description{
Identifies modules that are significantly correlated with traits.
}
\keyword{internal}
