% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_analysis.R
\name{load_dorothea_regulons}
\alias{load_dorothea_regulons}
\title{Load DoRothEA Regulons}
\usage{
load_dorothea_regulons(species, confidence_levels)
}
\arguments{
\item{species}{Species ("mouse" or "human")}

\item{confidence_levels}{Confidence levels to include}
}
\value{
DoRothEA regulon data frame
}
\description{
Loads DoRothEA transcription factor regulons.
}
\keyword{internal}
