% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_analysis.R
\name{map_regulons_to_ensembl}
\alias{map_regulons_to_ensembl}
\title{Map Regulons to ENSEMBL IDs}
\usage{
map_regulons_to_ensembl(regulon, species)
}
\arguments{
\item{regulon}{DoRothEA regulon data frame}

\item{species}{Species ("mouse" or "human")}
}
\value{
Regulon with ENSEMBL IDs
}
\description{
Maps gene symbols in regulons to ENSEMBL IDs using biomaRt.
}
\keyword{internal}
