% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution.R
\name{match_metadata_to_expression_immune}
\alias{match_metadata_to_expression_immune}
\title{Match Metadata to Immune Scores}
\usage{
match_metadata_to_expression_immune(
  immune_df,
  metadata,
  sample_id_col = "SampleID"
)
}
\arguments{
\item{immune_df}{Immune scores data frame}

\item{metadata}{Sample metadata}

\item{sample_id_col}{Sample ID column name}
}
\value{
Matched metadata
}
\description{
Helper function to match metadata with immune deconvolution scores.
}
\keyword{internal}
