% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{RNAseqAnalyzer}
\alias{-package}
\alias{RNAseqAnalyzer}
\alias{.onAttach}
\title{RNAseqAnalyzer: Comprehensive RNA-seq Analysis Pipeline}
\usage{
.onAttach(libname, pkgname)
}
\description{
This package provides functions for differential expression analysis,
pathway analysis, WGCNA co-expression networks, and visualization of RNA-seq data.
}
