% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.R
\name{run_camera_analysis}
\alias{run_camera_analysis}
\title{Run Camera Gene Set Testing}
\usage{
run_camera_analysis(
  voom_object,
  design_matrix,
  contrast_vector,
  gene_sets = NULL,
  species = "Mus musculus",
  min_size = 15
)
}
\arguments{
\item{voom_object}{Voom-transformed expression object}

\item{design_matrix}{Design matrix}

\item{contrast_vector}{Contrast vector}

\item{gene_sets}{List of gene sets (indices)}

\item{species}{Species for gene sets ("Mus musculus" or "Homo sapiens")}

\item{min_size}{Minimum gene set size (default: 15)}
}
\value{
Camera results data frame
}
\description{
Performs competitive gene set testing using limma's camera method.
}
