% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{save_gene_lists}
\alias{save_gene_lists}
\title{Save Gene Lists}
\usage{
save_gene_lists(
  efit,
  experiment_name,
  output_dir,
  fdr_threshold,
  lfc_threshold
)
}
\arguments{
\item{efit}{eBayes fit object from limma}

\item{experiment_name}{Name of the experiment}

\item{output_dir}{Output directory}

\item{fdr_threshold}{FDR threshold for significance}

\item{lfc_threshold}{Log fold change threshold}
}
\value{
None (invisible)
}
\description{
Saves lists of up- and down-regulated genes to files.
}
\keyword{internal}
